+++ CHANGELOG.md
@@ -8,6 +8,12 @@ The following allowances for breaking changes exist _for now_:
 
 TODO: Adopt more mature versioning, eventually. Read about eviction, binary compatibility, etc.
 
+#### v0.9 – Sep 2018
+
+* **API: `AriaAttrs` now exposes attributes as members, without `aria` object**
+* **API: Move `cls`, `className`, `rel`, `styleAttr`, and `textContent` into new `ComplexHtmlKeys` and `ComplexSvgKeys` traits, allowing for more freedom to define these keys in non-standard ways. Use `CanonicalComplexHtmlKeys` and `CanonicalComplexSvgKeys` to retain v0.8 functionality**
+* **API: remove `classNames` reflected attribute. You can define it locally or use the new `ComplexHtmlKeys` functionality to get the desired API**
+
 #### v0.8.1 – Sep 2018
 
 * **Fix: `span`, `fontSizeAdjust`, `listStyleImage.none` should be `lazy val`s, not `def`s**